pyproject.toml
[project]
name = "image-mosaic-generator"
version = "0.1.0"
description = "Create mosaics of images using a collection of smaller images"
authors = [
    { name = "Angelos Panagiotopoulos", email = "angelospanag@protonmail.com" },
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "scikit-learn>=1.6.1",
    "typer>=0.15.2",
]

[dependency-groups]
dev = ["ruff>=0.11.2"]
